openvibe-plugins-simple-visualisation :  * cmake now uses cmake-modules to find software components  + graz visualisation : added "show instruction" and "show feedback" check boxes  + graz visualisation : added double arrow when no instruction should be shown (this is equivalent to "do what you want")  + graz visualisation : estimated ratio of fail success over time windows  * graz visualisation : softened input value over several past buffers to avoid shaking  * spectrum display : added look up table for colors  * spectrum display : shows 10 seconds as default
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@701 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-plugins/simple-visualisation/trunc/src/ovpCSignalDisplay/ovpCChannelDisplayWidget.cpp
@@ -453,7 +453,7 @@ void CChannelDisplay::drawSignal()
 			float64 l_f64X = ((float64)((it->first-l_ui64StartTime)>>24) / (float64)((l_ui64EndTime-l_ui64StartTime)>>24)); //  ((l_i64BaseX + (j*l_f64WidthPerBuffer) + (i*l_f64PointStep) - m_f64TranslateX ) * m_f64ScaleX) ;
 			l_f64X = ((l_f64X - m_f64TranslateX)*m_f64ScaleX);
 			l_f64X *= l_iWidth;
-			gdk_draw_line(m_pDrawingArea->window, m_pDrawingArea->style->fg_gc[GTK_WIDGET_STATE (m_pDrawingArea)], l_f64X, 0, l_f64X, l_iHeight);
+			gdk_draw_line(m_pDrawingArea->window, m_pDrawingArea->style->fg_gc[GTK_WIDGET_STATE (m_pDrawingArea)], (int)l_f64X, 0, (int)l_f64X, l_iHeight);
 		}
 	}
